+++ Mintegral/MintegralAdapterConfiguration.h
@@ -36,8 +36,6 @@ extern NSString *const kMintegralErrorDomain;
 
 +(void)sdkInitialized;
 
-+(void)setGDPRInfo:(NSDictionary *)info;
-
 +(BOOL)isMute;
 
 +(void)setMute:(BOOL)muteStatus;